feat(#4) - 강의 검색 기능 추가 및 TestCode 추가
