+++ src/main/resources/com/flab/infrun/lecture/infrastructure/persistence/mybatis/mapper/LectureVideoFileMyBatisMapper.xml
@@ -2,40 +2,17 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Config 3.0//EN"
   "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper
-  namespace="com.flab.infrun.lecture.infrastructure.persistance.mybatis.mapper.LectureMyBatisMapper">
+  namespace="com.flab.infrun.lecture.infrastructure.persistence.mybatis.mapper.LectureVideoFileMyBatisMapper">
 
-  <select id="findById" resultType="com.flab.infrun.lecture.domain.Lecture">
-    SELECT id, name, nvl(price,0), nvl(lecture_level,0), skill, introduce, nvl(user_id,0)
-    FROM LECTURE
+  <select id="findById" resultType="com.flab.infrun.lecture.domain.LectureVideoFile">
+    SELECT ID , URL, NAME
+    FROM FILE
     WHERE ID = #{id}
   </select>
 
-  <select id="findByDetail" resultType="com.flab.infrun.lecture.domain.Lecture">
-    SELECT id , name, nvl(price,0), nvl(lecture_level,0), skill, introduce, nvl(user_id,0)
-    FROM LECTURE
-    <where>
-      <if test="lectureName !=null and lectureName != ''">
-        AND name = #{lectureName}
-      </if>
-      <if test="teacherName !=null and teacherName != ''">
-        AND user_name = #{teacherName}
-      </if>
-      <if test="price !=null and price != ''">
-        AND price = #{price}
-      </if>
-      <if test="lectureLevel !=null and lectureLevel != ''">
-        AND lecture_Level = #{lectureLevel}
-      </if>
-      <if test="skill !=null and skill != ''">
-        AND skill = #{skill}
-      </if>
-    </where>
-
-  </select>
-
   <insert id="save" useGeneratedKeys="true" keyProperty="id">
-    INSERT INTO LECTURE (NAME, PRICE, LECTURE_LEVEL, SKILL, INTRODUCE, USER_ID)
-    VALUES (#{name}, #{price}, #{lectureLevel}, #{skill}, #{introduce}, #{userId})
+    INSERT INTO FILE (URL, NAME)
+    VALUES (#{url}, #{name})
   </insert>
 
 </mapper>
